+++ CHANGELOG.md
@@ -7,40 +7,53 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 1.2.3 - 2022-12-15
+### Added
 
-## 1.2.2 - 2022-12-15
+- Icon button
 
 ## 1.2.1 - 2022-12-15
+
 ### Fixed
+
 - Refactor Alert
 
 ## 1.2.0 - 2022-12-15
+
 ### Added
+
 - Alert component
 - Storybook for Alert in 8 different states
 
 ### Fixed
+
 - Wrong class names on Toggle
 
 ## 1.1.9 - 2022-12-14
 
 ## 1.1.8 - 2022-12-14
+
 ### Fixed
+
 - Add peerDependencies for @tailwindcss/forms
 - Add name to Input-component
 - Add name to Textarea-component
 
 ## 1.1.7 - 2022-12-14
+
 ### Fixed
+
 - Updated classes on Input-component
 - Updated classes on Textarea-component
 
 ## 1.1.6 - 2022-12-14
+
 ### Fixed
+
 - Export InputType
 
 ## 1.1.3 - 2022-12-14
+
 ### Added
+
 - Changelog
 - An icon props on Input-component